requirements.txt
python-gnome2
